- Refactor UserChatBubble into a new file for better organization
- Use Medium font weight for User message
